Copy preferences.tsx into details-helper.tsx
For the instance menu to include the "collapse all instances" and
"expand all instances" buttons, the menu needs code from
details-helper.tsx. This creates an import loop that I couldn't easily
break, and so I'm merging the files. They're very closely related, and
no longer easy separate, so this doesn't seem like a bad thing to do.
